RIA-7359 RIA-7678 Notification refactoring and config
