+++ src/main/java/uk/gov/hmcts/reform/iacasenotificationsapi/infrastructure/config/NotificationHandlerConfiguration.java
@@ -2500,8 +2500,8 @@ public class NotificationHandlerConfiguration {
     }
 
     @Bean
-    public PreSubmitCallbackHandler<AsylumCase> internalAdaRespondentFtpaApplicationNotificationHandler(
-        @Qualifier("internalAdaRespondentFtpaApplicationNotificationGenerator")
+    public PreSubmitCallbackHandler<AsylumCase> internalRespondentFtpaApplicationNotificationHandler(
+        @Qualifier("internalRespondentFtpaApplicationNotificationGenerator")
         List<NotificationGenerator> notificationGenerators) {
 
         return new NotificationHandler(
@@ -2524,7 +2524,7 @@ public class NotificationHandlerConfiguration {
                     || callback.getEvent() == Event.RESIDENT_JUDGE_FTPA_DECISION)
                     && isRespondentApplication
                     && validDecisionOutcomeTypes.contains(decisionOutcomeType)
-                    && AsylumCaseUtils.isAcceleratedDetainedAppeal(asylumCase)
+                    && AsylumCaseUtils.isAppellantInDetention(asylumCase)
                     && AsylumCaseUtils.isInternalCase(asylumCase);
             },
             notificationGenerators
