+++ src/test/java/uk/gov/hmcts/reform/iacasenotificationsapi/domain/personalisation/respondent/RespondentNonStandardDirectionOfAppellantPersonalizationTest.java
@@ -5,9 +5,21 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.*;
+import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.APPEAL_REFERENCE_NUMBER;
+import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.APPELLANT_FAMILY_NAME;
+import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.APPELLANT_GIVEN_NAMES;
+import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.ARIA_LISTING_REFERENCE;
+import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.HEARING_CENTRE;
+import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.HOME_OFFICE_REFERENCE_NUMBER;
+import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.LIST_CASE_HEARING_CENTRE;
+import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.respondent.RespondentNonStandardDirectionOfAppellantPersonalization.CURRENT_CASE_STATE_VISIBLE_TO_HOME_OFFICE_ALL_FLAG_IS_NOT_PRESENT;
+import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.respondent.RespondentNonStandardDirectionOfAppellantPersonalization.EVENT_NOT_AVAILABLE;
 
-import java.util.*;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -50,8 +62,6 @@ class RespondentNonStandardDirectionOfAppellantPersonalizationTest {
     private final String homeOfficeHearingCentreEmail = "hc-taylorhouse@example.com";
     private final String homeOfficeEmail = "ho-taylorhouse@example.com";
     private final String homeOfficeFtpaEmailAddress = "ho-ftpa-taylorhouse@example.com";
-    private final String currentCaseStateVisibleToHomeOfficeAllFlagIsNotPresent = "currentCaseStateVisibleToHomeOfficeAll flag is not present";
-    private final String eventNotAvailable = "Send direction event not available in current state";
 
     private String directionDueDate = "2019-08-27";
     private String expectedDirectionDueDate = "27 Aug 2019";
@@ -70,7 +80,7 @@ class RespondentNonStandardDirectionOfAppellantPersonalizationTest {
     public void setUp() {
 
         when((emailAddressFinder.getListCaseHomeOfficeEmailAddress(asylumCase)))
-                .thenReturn(homeOfficeHearingCentreEmail);
+            .thenReturn(homeOfficeHearingCentreEmail);
 
         when(emailAddressFinder.getListCaseFtpaHomeOfficeEmailAddress(asylumCase)).thenReturn(homeOfficeFtpaEmailAddress);
 
@@ -90,14 +100,14 @@ class RespondentNonStandardDirectionOfAppellantPersonalizationTest {
 
 
         respondentNonStandardDirectionOfAppellantPersonalization = new RespondentNonStandardDirectionOfAppellantPersonalization(
-                templateId,
-                iaExUiFrontendUrl,
-                apcHomeOfficeEmailAddress,
-                lartHomeOfficeEmailAddress,
-                directionFinder,
-                customerServicesProvider,
-                appealService,
-                emailAddressFinder
+            templateId,
+            iaExUiFrontendUrl,
+            apcHomeOfficeEmailAddress,
+            lartHomeOfficeEmailAddress,
+            directionFinder,
+            customerServicesProvider,
+            appealService,
+            emailAddressFinder
         );
     }
 
@@ -110,32 +120,32 @@ class RespondentNonStandardDirectionOfAppellantPersonalizationTest {
     void should_return_correct_email_address_for_home_office() {
 
         List<State> apcEmail = newArrayList(
-                State.APPEAL_SUBMITTED,
-                State.PENDING_PAYMENT,
-                State.AWAITING_RESPONDENT_EVIDENCE,
-                State.AWAITING_CLARIFYING_QUESTIONS_ANSWERS,
-                State.CLARIFYING_QUESTIONS_ANSWERS_SUBMITTED
+            State.APPEAL_SUBMITTED,
+            State.PENDING_PAYMENT,
+            State.AWAITING_RESPONDENT_EVIDENCE,
+            State.AWAITING_CLARIFYING_QUESTIONS_ANSWERS,
+            State.CLARIFYING_QUESTIONS_ANSWERS_SUBMITTED
         );
 
 
         List<State> lartEmail = newArrayList(
-                State.CASE_UNDER_REVIEW,
-                State.RESPONDENT_REVIEW,
-                State.AWAITING_REASONS_FOR_APPEAL,
-                State.REASONS_FOR_APPEAL_SUBMITTED
+            State.CASE_UNDER_REVIEW,
+            State.RESPONDENT_REVIEW,
+            State.AWAITING_REASONS_FOR_APPEAL,
+            State.REASONS_FOR_APPEAL_SUBMITTED
         );
 
         List<State> pouNoListedEmail = newArrayList(
-                State.LISTING,
-                State.SUBMIT_HEARING_REQUIREMENTS
+            State.LISTING,
+            State.SUBMIT_HEARING_REQUIREMENTS
         );
 
         List<State> poulistedEmail = newArrayList(
-                State.PREPARE_FOR_HEARING,
-                State.FINAL_BUNDLING,
-                State.PRE_HEARING,
-                State.DECISION,
-                State.ADJOURNED
+            State.PREPARE_FOR_HEARING,
+            State.FINAL_BUNDLING,
+            State.PRE_HEARING,
+            State.DECISION,
+            State.ADJOURNED
         );
 
         Map<String, List<State>> states = new HashMap<>();
@@ -152,17 +162,17 @@ class RespondentNonStandardDirectionOfAppellantPersonalizationTest {
             List<State> statesList = states.get(emailAddress);
             for (State state : statesList) {
                 when(asylumCase.read(AsylumCaseDefinition.CURRENT_CASE_STATE_VISIBLE_TO_HOME_OFFICE_ALL, State.class))
-                        .thenReturn(Optional.of(state));
+                    .thenReturn(Optional.of(state));
 
                 if (emailAddress != null && emailAddress.equals(homeOfficeHearingCentreEmail)) {
                     // test the same state when the case is listed
                     when(appealService.isAppealListed(asylumCase)).thenReturn(true);
                     when(asylumCase.read(HEARING_CENTRE)).thenReturn(Optional.of(Optional.empty()));
                     when(asylumCase.read(LIST_CASE_HEARING_CENTRE, HearingCentre.class))
-                            .thenReturn(Optional.of(HearingCentre.TAYLOR_HOUSE));
+                        .thenReturn(Optional.of(HearingCentre.TAYLOR_HOUSE));
 
                     assertTrue(respondentNonStandardDirectionOfAppellantPersonalization.getRecipientsList(asylumCase)
-                            .contains(homeOfficeHearingCentreEmail));
+                        .contains(homeOfficeHearingCentreEmail));
                 } else if (emailAddress != null && emailAddress.equals(homeOfficeEmail)) {
                     //case not listed yet
                     when(appealService.isAppealListed(asylumCase)).thenReturn(false);
@@ -170,10 +180,10 @@ class RespondentNonStandardDirectionOfAppellantPersonalizationTest {
                     when(asylumCase.read(LIST_CASE_HEARING_CENTRE, HearingCentre.class)).thenReturn(Optional.empty());
 
                     assertTrue(respondentNonStandardDirectionOfAppellantPersonalization.getRecipientsList(asylumCase)
-                            .contains(homeOfficeEmail));
+                        .contains(homeOfficeEmail));
                 } else {
                     assertTrue(respondentNonStandardDirectionOfAppellantPersonalization.getRecipientsList(asylumCase)
-                            .contains(emailAddress));
+                        .contains(emailAddress));
                 }
             }
         }
@@ -182,11 +192,11 @@ class RespondentNonStandardDirectionOfAppellantPersonalizationTest {
     @Test
     void should_throw_exception_if_current_visible_state_to_home_office_all_is_not_present() {
         when(asylumCase.read(AsylumCaseDefinition.CURRENT_CASE_STATE_VISIBLE_TO_HOME_OFFICE_ALL, State.class))
-                .thenReturn(Optional.empty());
+            .thenReturn(Optional.empty());
 
         assertThatThrownBy(() -> respondentNonStandardDirectionOfAppellantPersonalization.getRecipientsList(asylumCase))
-                .isExactlyInstanceOf(IllegalStateException.class)
-                .hasMessage(currentCaseStateVisibleToHomeOfficeAllFlagIsNotPresent);
+            .isExactlyInstanceOf(IllegalStateException.class)
+            .hasMessage(CURRENT_CASE_STATE_VISIBLE_TO_HOME_OFFICE_ALL_FLAG_IS_NOT_PRESENT);
     }
 
     @ParameterizedTest
@@ -195,32 +205,32 @@ class RespondentNonStandardDirectionOfAppellantPersonalizationTest {
     })
     void should_throw_exception_when_send_direction_event_not_available_in_current_state(State currentState) {
         when(asylumCase.read(AsylumCaseDefinition.CURRENT_CASE_STATE_VISIBLE_TO_HOME_OFFICE_ALL, State.class))
-                .thenReturn(Optional.of(currentState));
+            .thenReturn(Optional.of(currentState));
 
         assertThatThrownBy(() -> respondentNonStandardDirectionOfAppellantPersonalization.getRecipientsList(asylumCase))
-                .isExactlyInstanceOf(IllegalStateException.class)
-                .hasMessage(eventNotAvailable);
+            .isExactlyInstanceOf(IllegalStateException.class)
+            .hasMessage(EVENT_NOT_AVAILABLE);
     }
 
     @Test
     void should_return_given_reference_id() {
         assertEquals(caseId + "_RESPONDENT_NON_STANDARD_DIRECTION_OF_APPELLANT",
-                respondentNonStandardDirectionOfAppellantPersonalization.getReferenceId(caseId));
+            respondentNonStandardDirectionOfAppellantPersonalization.getReferenceId(caseId));
     }
 
     @Test
     public void should_throw_exception_on_personalisation_when_case_is_null() {
 
         assertThatThrownBy(() -> respondentNonStandardDirectionOfAppellantPersonalization.getPersonalisation((AsylumCase) null))
-                .isExactlyInstanceOf(NullPointerException.class)
-                .hasMessage("asylumCase must not be null");
+            .isExactlyInstanceOf(NullPointerException.class)
+            .hasMessage("asylumCase must not be null");
     }
 
     @Test
     public void should_return_personalisation_when_all_information_given() {
 
         Map<String, String> personalisation =
-                respondentNonStandardDirectionOfAppellantPersonalization.getPersonalisation(asylumCase);
+            respondentNonStandardDirectionOfAppellantPersonalization.getPersonalisation(asylumCase);
         String listingReferenceLine = "\nListing reference: " + ariaListingReference;
 
         assertEquals(appealReferenceNumber, personalisation.get("appealReferenceNumber"));
@@ -245,7 +255,7 @@ class RespondentNonStandardDirectionOfAppellantPersonalizationTest {
         when(asylumCase.read(HOME_OFFICE_REFERENCE_NUMBER, String.class)).thenReturn(Optional.empty());
 
         Map<String, String> personalisation =
-                respondentNonStandardDirectionOfAppellantPersonalization.getPersonalisation(asylumCase);
+            respondentNonStandardDirectionOfAppellantPersonalization.getPersonalisation(asylumCase);
 
         assertEquals("", personalisation.get("appealReferenceNumber"));
         assertEquals("", personalisation.get("listingReferenceLine"));
@@ -264,7 +274,7 @@ class RespondentNonStandardDirectionOfAppellantPersonalizationTest {
         when(directionFinder.findFirst(asylumCase, DirectionTag.NONE)).thenReturn(Optional.empty());
 
         assertThatThrownBy(() -> respondentNonStandardDirectionOfAppellantPersonalization.getPersonalisation(asylumCase))
-                .isExactlyInstanceOf(IllegalStateException.class)
-                .hasMessage("non-standard direction is not present");
+            .isExactlyInstanceOf(IllegalStateException.class)
+            .hasMessage("non-standard direction is not present");
     }
 }
